Split up styles into seperate files
