+++ frontend/src/components/Footer/footerStyle.js
@@ -1,4 +1,3 @@
-import { Link } from 'react-router-dom'
 import { styled } from 'goober'
 
 export const Container = styled('footer')`
