+++ frontend/src/pages/NewFile/components/CardList/CardList.js
@@ -1,4 +1,4 @@
-import { styled } from 'goober'
+import { CardListContainer, CardListTitleContainer, CardList } from './cardListStyle'
 
 const CardList = ({ title, children, button=null }) => <>
   <CardListTitleContainer>
@@ -10,22 +10,5 @@ const CardList = ({ title, children, button=null }) => <>
   </CardListContainer>
 </>
 
-const CardListTitleContainer = styled('div')`
-  display: flex;
-  width: 100%;
-  gap: 1em;
-  align-items: center;
-`
-
-const CardListTitle = styled('h2')`
-  margin-block: 0;
-`
-
-const CardListContainer = styled('div')`
-  display: flex;
-  gap: .4em;
-  margin-block-start: 1em;
-  margin-block-end: 3em;
-`
 
 export default CardList
