+++ frontend/src/pages/NewFile/components/CardList/cardListStyle.js
@@ -1,0 +1,19 @@
+import { styled } from 'goober'
+
+export const CardListTitleContainer = styled('div')`
+  display: flex;
+  width: 100%;
+  gap: 1em;
+  align-items: center;
+`
+
+export const CardListTitle = styled('h2')`
+  margin-block: 0;
+`
+
+export const CardListContainer = styled('div')`
+  display: flex;
+  gap: .4em;
+  margin-block-start: 1em;
+  margin-block-end: 3em;
+`
