+++ frontend/src/pages/NewFile/components/NewProjectCard/NewProjectCard.js
@@ -1,4 +1,4 @@
-import { styled } from 'goober'
+import { CardContainer, CardImage, CardContent } from './newProjectCardStyle'
 
 const NewProjectCard = ({ title, description, disabled, ...props }) => {
   return <CardContainer $disabled={disabled} {...props}>
@@ -9,67 +9,5 @@ const NewProjectCard = ({ title, description, disabled, ...props }) => {
     </CardContent>
   </CardContainer>
 }
-
-const CardContainer = styled('div')`
-  display: grid;
-  grid-template-rows: 1.3fr 1fr;
-  box-sizing: border-box;
-  min-width: 15em;
-  width: 20em;
-
-  border-radius: .5rem;
-  border: 3px solid transparent;
-  overflow: hidden;
-
-  ${p => !p.$disabled && `
-    cursor: pointer;
-    &:hover {
-      border: 3px solid var(--primary);
-    }
-  `}
-
-  ${p => p.$disabled && `
-    opacity: .7; 
-  `}
-`
-
-const CardContent = styled('div')`
-  background: var(--toolbar);
-  padding: .9em;
-
-  > p { 
-    margin: 0;
-    margin-top: .5em;
-    font-size: .9rem;
-  }
-`
-
-const CardImage = styled('div')`
-  /* TODO: use image in place of temp background */
-  --dot-fraction: 12.5%;
-  background: radial-gradient(
-    var(--grid-dot),
-    var(--grid-dot) var(--dot-fraction),
-    var(--white) var(--dot-fraction));
-  background-size: 30px 30px;
-  background-position: 5px 5px;
-
-  ${p => p.$disabled && `
-    position: relative;
-    &::after {
-      content: 'Coming Soon';
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      color: grey;
-      font-size: 1.7rem;
-      opacity: .4;
-      position: absolute;
-      inset: 0;
-      font-family: var(--font-feature);
-    } 
-  `}
-`
-
 
 export default NewProjectCard
