+++ frontend/src/pages/NotFound/notFoundStyle.js
@@ -1,0 +1,15 @@
+import { styled } from 'goober'
+
+export const Code = styled('div')`
+  padding: 1em;
+  background: var(--toolbar);
+  font-family: monospace;
+  border-radius: .5rem;
+`
+
+export const Hint = styled('p')`
+  margin-top: 100vh;
+  color: grey;
+  opacity: .2;
+  border-radius: .5rem;
+`
